Extract mutation
